hr_portal BKP (14-FEB-2019)/flask-hr(12-FEB-2019)/requirements.txt
absl-py==0.7.0
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.6
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.4.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==2.3.1
cycler==0.10.0
cymem==2.0.2
Cython==0.28.2
dash==0.21.0
dash-core-components==0.22.1
dash-html-components==0.10.0
dash-renderer==0.12.1
dbConnect==2.1
decorator==4.3.0
dill==0.2.8.2
Django==1.8
django-registration-redux==2.5
docopt==0.6.2
dominate==2.3.4
fakeredis==0.10.3
fbmessenger==5.4.0
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.4.0
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.1.5
Flask-WTF==0.14.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
gmane==0.1.dev25
greenlet==0.4.15
grpcio==1.18.0
h5py==2.9.0
heroku==0.1.4
humanfriendly==4.18
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython==7.1.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10
jsonpickle==1.1
jsonschema==2.6.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mattermostwrapper==2.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
mysql==0.0.2
mysql-connector==2.1.6
mysql-connector-python==8.0.12
mysqlclient==1.4.2.post1
nano==0.9.4
nbformat==4.4.0
networkx==2.2
nltk==3.4
numpy==1.16.1
packaging==18.0
pandas==0.22.0
pandas-datareader==0.6.0
parso==0.3.1
passlib==1.7.1
pbr==5.1.1
pickleshare==0.7.5
pika==0.12.0
plac==0.9.6
plotly==2.5.1
preshed==2.0.1
prompt-toolkit==2.0.7
protobuf==3.6.1
PyAudio==0.2.11
pycparser==2.18
pydot==1.4.1
pygame==1.9.4
Pygments==2.2.0
pygorithm==1.0.4
PyHamcrest==1.9.0
PyJWT==1.7.1
pykwalify==1.7.0
PyMySQL==0.9.2
pyparsing==2.2.0
pyreadline==2.1
PySocks==1.6.8
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2018.9
PyYAML==3.13
rasa-core-sdk==0.12.2
redis==2.10.6
requests==2.21.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-toolbelt==0.9.1
ruamel.yaml==0.15.88
scikit-learn==0.19.2
selenium==3.141.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
slackclient==1.3.0
social==0.1.dev3
SpeechRecognition==3.8.1
splinter==0.10.0
SQLAlchemy==1.2.18
stevedore==1.30.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminaltables==3.1.0
toolz==0.9.0
tqdm==4.28.1
traitlets==4.3.2
twilio==6.24.1
typing==3.6.6
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.5
visitor==0.1.3
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.55.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
zope.interface==4.6.0
